commands: Own selected-change completion flow
Include, discard, and skip commands share post-action selection
handling that decides whether to advance, stay quiet, print an
empty-selection message, or display the next selected change.

That flow belongs to command orchestration rather than hunk navigation
state. Keeping it in hunk_tracking makes command follow-up behavior
look like a data module responsibility.

This commit addresses that by moving the completion helper into
commands.selection.action_completion, updating the command callers, and
removing the hunk_tracking copy.

The next commit will add an architecture guard for the command-owned
helper.
